% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions get data.R
\name{clear_data_acres}
\alias{clear_data_acres}
\title{Clear and Reformat Data for Acres}
\usage{
clear_data_acres(dat)
}
\arguments{
\item{dat}{A data frame containing the columns `year`, `county_name`, `state_fips_code`, `county_ansi`, `asd_desc`, `asd_code`, `domain_desc`, `domaincat_desc`, and `Value`. The `Value` column should contain the area in acres, potentially formatted as a string with commas.}
}
\value{
A cleaned and reformatted data frame with columns for Year, County_Name, Region, Region_code, Domain, Size.bins, and Irr.Area.Acres. The `fips` column is also added, representing the concatenated state and county codes. Columns for state and county codes are removed from the output.
}
\description{
This function cleans and reformats the input dataset for analysis. It replaces commas in the `Value` field, converts the `Value` field from UTF-8 to ASCII to remove all characters except digits, and selects specific columns for the final dataset. It also creates a new `fips` column by concatenating the state and county codes.
}
\examples{
# Assuming `df` is your dataset:
cleaned_data <- clear_data_acres(df)

}
