% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions get data.R
\name{clear_data_ops}
\alias{clear_data_ops}
\title{Clean Data for Operations}
\usage{
clear_data_ops(dat)
}
\arguments{
\item{dat}{A data frame containing the columns: `Value`, `year`, `county_name`, `state_fips_code`, `county_ansi`, `asd_desc`, `asd_code`, `domain_desc`, and `domaincat_desc`.}
}
\value{
A cleaned data frame with the columns: `Year`, `County_Name`, `Region`, `Region_code`, `Domain`, `Size.bins`, `Irr.Ops`, and `fips`.
}
\description{
This function cleans the agricultural land data related to the number of operations.
Similar to `clear_data_acres`, it processes the `Value` column by removing commas and converting from UTF-8 to ASCII,
selects and renames relevant columns for clarity, creates a `fips` column by combining state and county codes, and
drops the original state and county code columns.
}
\examples{
# Assuming `dat` is your dataframe:
cleaned_ops_data <- clear_data_ops(dat)
}
