% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions get data.R
\name{get_nass_data}
\alias{get_nass_data}
\title{Fetch and Process NASS Data for a Given Year and State with API Key}
\usage{
get_nass_data(Year, state_name, my_nass_key)
}
\arguments{
\item{Year}{A numeric value representing the year for which the data is to be fetched.}

\item{state_name}{A character string specifying the name of the state for which the data is to be fetched.}

\item{my_nass_key}{A character string representing the NASS API key required for accessing the data.}
}
\value{
A data frame containing the merged and cleaned dataset of irrigated acres and number of operations
        for the specified year and state. If discrepancies are found, an empty data frame is returned with a warning.
}
\description{
This function sets the NASS API key, fetches agricultural land data related to irrigated acres and the number of operations
for a specified year and state, cleans, and merges these datasets. The function ensures that there
are no discrepancies between the datasets before merging. If discrepancies are found, it returns an empty
data frame with a warning message. Otherwise, it returns a cleaned and merged dataset.
}
\examples{
# Assuming nass_data, clear_data_acres, and clear_data_ops functions are defined and available,
# and you have a valid NASS API key:
my_api_key <- "your_nass_api_key_here"
nass_data_2022_SC <- get_nass_data(2022, "South Carolina", my_api_key)
}
